% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_utilities.R
\name{animate_plot}
\alias{animate_plot}
\title{Generate animated gif from set of static images}
\usage{
animate_plot(file_names, output_file, delay = 125)
}
\arguments{
\item{file_names}{A vector of file names for the image files}

\item{output_file}{File name/path for output}

\item{delay}{A number that measures the delay in 1/100 seconds.
Default is 125.}
}
\value{
Returns animated gif saved as output_file
}
\description{
Generate animated gif from set of static images
}
\examples{
animate_plot(c("image1.png", "image2.png", "image3.png"), "animated_image.gif", 100)
}
