% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_utilities.R
\name{convert_svg_pdf}
\alias{convert_svg_pdf}
\title{Generate PDF file from grViZ object}
\usage{
convert_svg_pdf(gv, output_file)
}
\arguments{
\item{gv}{An object of type grViz from DiagrammeR package}

\item{output_file}{File name/path for pdf}
}
\value{
Returns a pdf saved as output_file
}
\description{
Generate PDF file from grViZ object
}
\examples{
library(dm)
a <- dm(mtcars, iris) \%>\% dm_draw(rankdir = "LR", view_type = "all")
convert_svg_pdf(a, "./example.pdf")
}
