% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimgeo_fx.R
\name{lonlat_to_geoid}
\alias{lonlat_to_geoid}
\title{Function for Lon/Lat -> GEOID (General)}
\usage{
lonlat_to_geoid(pointsDF, geo, name_col = "GEOID")
}
\arguments{
\item{pointsDF}{A data.frame whose first column contains longitudes and}

\item{geo}{A submitted geospatial polygon for census block groups}

\item{name_col}{Name of a column in geo that supplies data point to return}
}
\value{
Returns geography names
}
\description{
Function for Lon/Lat -> GEOID (General)
}
\examples{
lonlat_to_geoid(df, geo = tigris::states())
}
