% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimgeo_fx.R
\name{round_lat_long}
\alias{round_lat_long}
\title{Function for getting the latitude and longitude rounded to proper level for use with dim_geo objects}
\usage{
round_lat_long(input, dim_geo_level)
}
\arguments{
\item{input}{A longitude or latitude value to be rounded}

\item{dim_geo_level}{The level of rounding to use}
}
\value{
Returns rounded longitude/latitude value
}
\description{
Function for getting the latitude and longitude rounded to proper level for use with dim_geo objects
}
\examples{
round_lat_long(-84.82549, 4)
}
